Integrate XYZ, XYM, XYZM tests for geoarrow-array (#1046)
### Change list

- Leave only the raw geometries in the `geoarrow-test` crate
- Move test array definitions into `geoarrow-array/test`
- At least add test for converting to/from `arrow` representations with
all dimensions of geometry data
- Fix coordinate buffer builder with xyzm dimension

+++ rust/geoarrow-test/src/raw/point.rs
@@ -3,10 +3,10 @@ use wkt::wkt;
 pub mod xy {
     use super::*;
 
-    pub fn geoms() -> Vec<Option<wkt::types::Point<i32>>> {
+    pub fn geoms() -> Vec<Option<wkt::types::Point<f64>>> {
         vec![
-            Some(wkt! { POINT (30 10) }),
-            Some(wkt! { POINT (40 20) }),
+            Some(wkt! { POINT (30. 10.) }),
+            Some(wkt! { POINT (40. 20.) }),
             None,
             Some(wkt! { POINT EMPTY }),
         ]
@@ -16,10 +16,10 @@ pub mod xy {
 pub mod xyz {
     use super::*;
 
-    pub fn geoms() -> Vec<Option<wkt::types::Point<i32>>> {
+    pub fn geoms() -> Vec<Option<wkt::types::Point<f64>>> {
         vec![
-            Some(wkt! { POINT Z (30 10 40) }),
-            Some(wkt! { POINT Z (40 20 60) }),
+            Some(wkt! { POINT Z (30. 10. 40.) }),
+            Some(wkt! { POINT Z (40. 20. 60.) }),
             None,
             Some(wkt! { POINT Z EMPTY }),
         ]
@@ -29,10 +29,10 @@ pub mod xyz {
 pub mod xym {
     use super::*;
 
-    pub fn geoms() -> Vec<Option<wkt::types::Point<i32>>> {
+    pub fn geoms() -> Vec<Option<wkt::types::Point<f64>>> {
         vec![
-            Some(wkt! { POINT M (30 10 300) }),
-            Some(wkt! { POINT M (40 20 800) }),
+            Some(wkt! { POINT M (30. 10. 300.) }),
+            Some(wkt! { POINT M (40. 20. 800.) }),
             None,
             Some(wkt! { POINT M EMPTY }),
         ]
@@ -42,10 +42,10 @@ pub mod xym {
 pub mod xyzm {
     use super::*;
 
-    pub fn geoms() -> Vec<Option<wkt::types::Point<i32>>> {
+    pub fn geoms() -> Vec<Option<wkt::types::Point<f64>>> {
         vec![
-            Some(wkt! { POINT ZM (30 10 40 300) }),
-            Some(wkt! { POINT ZM (40 20 60 800) }),
+            Some(wkt! { POINT ZM (30. 10. 40. 300.) }),
+            Some(wkt! { POINT ZM (40. 20. 60. 800.) }),
             None,
             Some(wkt! { POINT ZM EMPTY }),
         ]
